requirements.txt
discord.py
asyncpg==0.21.0
